evlog: extract dataProviderCapture helper from remultDrains
